pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "note-research-team"
version = "0.1.0"
description = "CrewAI note research team with Notion integration"
requires-python = ">=3.8"
dependencies = [
    "crewai>=0.22.0",
    "crewai-tools>=0.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["src"]
